+++ src/main/java/com/atlauncher/data/Account.java
@@ -27,15 +27,27 @@ import com.atlauncher.gui.tabs.InstancesTab;
 import com.atlauncher.gui.tabs.PacksTab;
 import com.atlauncher.utils.Authentication;
 import com.atlauncher.utils.HTMLUtils;
+import com.atlauncher.utils.MojangAPIUtils;
 import com.atlauncher.utils.Utils;
 import com.mojang.authlib.yggdrasil.YggdrasilUserAuthentication;
 import com.mojang.util.UUIDTypeAdapter;
 
 import javax.imageio.ImageIO;
-import javax.swing.*;
-import java.awt.*;
+import javax.swing.ImageIcon;
+import javax.swing.JLabel;
+import javax.swing.JOptionPane;
+import javax.swing.JPanel;
+import javax.swing.JPasswordField;
+import java.awt.BorderLayout;
+import java.awt.Graphics;
+import java.awt.Image;
 import java.awt.image.BufferedImage;
-import java.io.*;
+import java.io.BufferedReader;
+import java.io.File;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.io.ObjectInputStream;
+import java.io.Serializable;
 import java.net.HttpURLConnection;
 import java.net.MalformedURLException;
 import java.net.URL;
@@ -298,6 +310,10 @@ public class Account implements Serializable {
      */
     public String getUUID() {
         return (this.uuid == null ? "0" : this.uuid);
+    }
+
+    public boolean isUUIDNull() {
+        return this.uuid == null;
     }
 
     /**
@@ -649,5 +665,27 @@ public class Account implements Serializable {
         }
 
         return response;
+    }
+
+    public boolean checkForUsernameChange() {
+        if (this.uuid == null) {
+            LogManager.error("The account " + this.minecraftUsername + " has no UUID associated with it !");
+            return false;
+        }
+
+        String currentUsername = MojangAPIUtils.getCurrentUsername(this.getUUIDNoDashes());
+
+        if (currentUsername == null) {
+            return false;
+        }
+
+        if (!currentUsername.equals(this.minecraftUsername)) {
+            LogManager.info("The username for account with UUID of " + this.getUUIDNoDashes() + " changed from " +
+                    this.minecraftUsername + " to " + currentUsername);
+            this.minecraftUsername = currentUsername;
+            return true;
+        }
+
+        return false;
     }
 }