+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog
 ====================================
 
-### 3.2.2.7
+### 3.2.3.0
 
--
+- Fix issues with filesizes not showing correctly when downloading mods for a pack
+- Fix issues with username changes causing users to not be able to login with Mojang anymore
+- Add in --force-offline-mode command line argument to, well, force offline mode. Should work fine
+- Fix semi public packs being added/removed not refreshing the packs list
+- Change the Install Just For Me to Enable User Lock and add in a warning popup
+- Change the User Lock option to use only UUID's so that they aren't affected by username changes